work buttons

+++ backend/src/friends/friends.service.ts
@@ -88,37 +88,58 @@ export class FriendsService {
         },
       },
     });
-    if (uniqueBlock) {
-      if (friendshipUserToFriend && friendshipUserToFriend.status == 3) {
-        await this.prisma.friendship.update({
-          where: {
-            userId_friendId: {
-              userId: userId,
-              friendId: friendId,
-            },
+    // if (uniqueBlock) {
+    //   if (friendshipUserToFriend && friendshipUserToFriend.status == 3) {
+    //     await this.prisma.friendship.update({
+    //       where: {
+    //         userId_friendId: {
+    //           userId: userId,
+    //           friendId: friendId,
+    //         },
+    //       },
+    //       data: {
+    //         status: 2,
+    //       },
+    //     });
+    //     await this.prisma.blocked.delete({
+    //       where: {
+    //         userId_blockedId: {
+    //           userId: userId,
+    //           blockedId: friendId,
+    //         },
+    //       },
+    //     });
+    //   } else if (!friendshipUserToFriend) {
+    //     await this.prisma.blocked.delete({
+    //       where: {
+    //         userId_blockedId: {
+    //           userId: userId,
+    //           blockedId: friendId,
+    //         },
+    //       },
+    //     });
+    //   }
+    // }
+    if (uniqueBlock && friendshipUserToFriend.status == 3) {
+      await this.prisma.friendship.update({
+        where: {
+          userId_friendId: {
+            userId: userId,
+            friendId: friendId,
           },
-          data: {
-            status: 2,
+        },
+        data: {
+          status: 2,
+        },
+      });
+      await this.prisma.blocked.delete({
+        where: {
+          userId_blockedId: {
+            userId: userId,
+            blockedId: friendId,
           },
-        });
-        await this.prisma.blocked.delete({
-          where: {
-            userId_blockedId: {
-              userId: userId,
-              blockedId: friendId,
-            },
-          },
-        });
-      } else if (!friendshipUserToFriend) {
-        await this.prisma.blocked.delete({
-          where: {
-            userId_blockedId: {
-              userId: userId,
-              blockedId: friendId,
-            },
-          },
-        });
-      }
+        },
+      });
     }
     return this.getFriendsList(userId);
   }
@@ -149,34 +170,54 @@ export class FriendsService {
         },
       },
     });
-    if (!uniqueBlock) {
-      if (friendshipUserToFriend && friendshipUserToFriend.status !== 3) {
-        await this.prisma.friendship.update({
-          where: {
-            userId_friendId: {
-              userId: userId,
-              friendId: friendId,
-            },
-          },
-          data: {
-            status: 3,
-          },
-        });
-        await this.prisma.blocked.create({
-          data: {
+    // if (!uniqueBlock) {
+    //   if (friendshipUserToFriend && friendshipUserToFriend.status !== 3) {
+    //     await this.prisma.friendship.update({
+    //       where: {
+    //         userId_friendId: {
+    //           userId: userId,
+    //           friendId: friendId,
+    //         },
+    //       },
+    //       data: {
+    //         status: 3,
+    //       },
+    //     });
+    //     await this.prisma.blocked.create({
+    //       data: {
+    //         userId: userId,
+    //         blockedId: friendId,
+    //       },
+    //     });
+    //   } else if (!friendshipUserToFriend) {
+    //     await this.prisma.blocked.create({
+    //       data: {
+    //         userId: userId,
+    //         blockedId: friendId,
+    //       },
+    //     });
+    //   }
+    // }
+    if (!uniqueBlock && friendshipUserToFriend.status !== 3) {
+      await this.prisma.friendship.update({
+        where: {
+          userId_friendId: {
             userId: userId,
-            blockedId: friendId,
+            friendId: friendId,
           },
-        });
-      } else if (!friendshipUserToFriend) {
-        await this.prisma.blocked.create({
-          data: {
-            userId: userId,
-            blockedId: friendId,
-          },
-        });
-      }
+        },
+        data: {
+          status: 3,
+        },
+      });
+      await this.prisma.blocked.create({
+        data: {
+          userId: userId,
+          blockedId: friendId,
+        },
+      });
     }
+
     return this.getFriendsList(userId);
   }
